wavelet.cpp - changed double to float
git-svn-id: https://code.nr.no/scm/svn/crava/trunk@404 176de4d3-4747-0410-aa96-47f6d6634214

+++ src/wavelet.cpp
@@ -616,7 +616,7 @@ Wavelet::calculateSNRatioAndLocalWavelet(Simbox        * simbox,
 
     if(doEstimateLocalNoise)
     {
-      double errStdLN;
+      float errStdLN;
       if(doEstimateSNRatio==true)
         errStdLN = errStd;
       else //SNRatio given in model file
